+++ host-agent/src/handlers/mod.rs
@@ -1,3 +1,4 @@
 pub mod files;
+pub mod hosts;
 pub mod network;
 //pub mod process;